24 Commit

+++ 2_Trimestre/Java/Pruebas/src/main/java/Prueba3/Ej3.java
@@ -7,7 +7,7 @@ public class Ej3 {
 	
 	private static List<Integer> edades = new ArrayList<Integer>();
 
-	public static int obtenerPromedioEdades(List<Integer> edades) {
+	public static double obtenerPromedioEdades(List<Integer> edades) {
 		int num = 0;
 		
 		for(int edad : edades) {
